[level-wise-counts-exp] TreeGrammarStructure/TreeEncoder and WindSpeed data
Downloaded and preprocessed wind speed dataset for regression. Refactoring of TreeGrammarStructure and TreeEncoder with generation of encoding and scaler for each representation: counts, level-wise counts, and one-hot. Moreover, added methods for computing add_prop as ground truth for each of the aforemetnioned representations. Renamed nsgp.util in nsgp.structure. Added method in TreeGrammarStructure for computing linear model discovered in the math formula interpretability paper. Added method in TreeGrammarStructure for computing size of encoding for each of the aforementioned representations.